:herb: split server.vim tests
- Tests can now be run separately.
- Improved parallelism in CI.
